- Add delete function for Standard, Dynamic List, Chart, Page, App, Database - More ConfirmationOptions from ButtonOptions to ActionCommandOptions
