+++ src/web-apis/.editorconfig
@@ -143,3 +143,6 @@ dotnet_diagnostic.CA1054.severity = none
 
 # CA1707: Identifiers should not contain underscores
 dotnet_diagnostic.CA1707.severity = none
+
+# CA1062: Validate arguments of public methods
+dotnet_diagnostic.CA1062.severity = none
