+++ src/web-apis/LetPortal.Portal/Entities/Shared/ButtonOptions.cs
@@ -3,19 +3,10 @@
 namespace LetPortal.Portal.Entities.Shared
 {
     public class ButtonOptions
-    {
-        public ConfirmationOptions ConfirmationOptions { get; set; }
-
+    {                              
         public ActionCommandOptions ActionCommandOptions { get; set; }
 
         public RouteOptions RouteOptions { get; set; }
-    }
-
-    public class ConfirmationOptions
-    {
-        public bool IsEnable { get; set; }
-
-        public string ConfirmationText { get; set; }
     }
 
     public class ActionOptions
