+++ src/web-apis/LetPortal.Portal/Entities/Shared/ConfirmationOptions.cs
@@ -1,0 +1,9 @@
+﻿namespace LetPortal.Portal.Entities.Shared
+{
+    public class ConfirmationOptions
+    {
+        public bool IsEnable { get; set; }
+
+        public string ConfirmationText { get; set; }
+    }
+}
