+++ src/web-apis/LetPortal.Portal/Entities/Shared/NotificationOptions.cs
@@ -1,7 +1,7 @@
 ﻿namespace LetPortal.Portal.Entities.Shared
 {
     public class NotificationOptions
-    {
+    {        
         public string CompleteMessage { get; set; }
 
         public string FailedMessage { get; set; }
