+++ src/web-apis/LetPortal.Versions/Apps/App_0_1_0.cs
@@ -6,9 +6,9 @@ using LetPortal.Portal.Entities.Menus;
 
 namespace LetPortal.Versions.Apps
 {
-    public class App_0_0_1 : IPortalVersion
+    public class App_0_1_0 : IPortalVersion
     {
-        public string VersionNumber => "0.0.1";
+        public string VersionNumber => "0.1.0";
 
         public void Downgrade(IVersionContext versionContext)
         {
