+++ src/web-apis/LetPortal.Versions/Components/Charts/Chart_0_1_0.cs
@@ -2,11 +2,11 @@
 using LetPortal.Core.Versions;
 using LetPortal.Portal.Entities.Components;
 
-namespace LetPortal.Versions.Components
+namespace LetPortal.Versions.Components.Charts
 {
-    public class Chart_0_0_1 : IPortalVersion
+    public class Chart_0_1_0 : IPortalVersion
     {
-        public string VersionNumber => "0.0.1";
+        public string VersionNumber => "0.1.0";
 
         public void Downgrade(IVersionContext versionContext)
         {
