+++ src/web-apis/LetPortal.Versions/Components/Standards/Standard_0_2_0.cs
@@ -4,23 +4,24 @@ using LetPortal.Portal.Entities.SectionParts;
 
 namespace LetPortal.Versions.Components
 {
-    public class Standard_0_0_1 : IPortalVersion
+    public class Standard_0_2_0 : IPortalVersion
     {
-        public string VersionNumber => "0.0.1";
+        public string VersionNumber => "0.2.0";
 
         public void Downgrade(IVersionContext versionContext)
         {
-            versionContext.DeleteData<StandardComponent>("5d25f6abe88d28422061cfaf");
-            versionContext.DeleteData<StandardComponent>("5d3836194d8fa90874135d68");
+            versionContext.DeleteData<StandardComponent>("5d615e501773a96ee8eb5ed3");
+            versionContext.DeleteData<StandardComponent>("5d6222ec3aae6a79ecf035e9");
+            versionContext.DeleteData<StandardComponent>("5d63423dbaac7d4790b7d300");
         }
 
         public void Upgrade(IVersionContext versionContext)
         {
-            var databaseStandard = new StandardComponent
+            var userFormStandard = new StandardComponent
             {
-                Id = "5d25f6abe88d28422061cfaf",
-                Name = "database-form",
-                DisplayName = "Database Form",
+                Id = "5d615e501773a96ee8eb5ed3",
+                Name = "user-form",
+                DisplayName = "User Form",
                 LayoutType = PageSectionLayoutType.OneColumn,
                 Controls = new List<Portal.Entities.SectionParts.Controls.PageControl>
                 {
@@ -60,20 +61,57 @@ namespace LetPortal.Versions.Components
                     },
                     new Portal.Entities.SectionParts.Controls.PageControl
                     {
-                        Name = "name",
+                        Name = "username",
                         Order = 1,
-                        Type = Portal.Entities.SectionParts.Controls.ControlType.Textbox,
+                        Type = Portal.Entities.SectionParts.Controls.ControlType.Label,
                         Options = new List<Portal.Entities.Pages.ShellOption>
                         {
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "label",
-                               Value = "Name"
+                               Value = "Username"
                            },
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "placeholder",
-                               Value = "Name"
+                               Value = "Enter a username"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "disabled",
+                               Value = "!!queryparams.id"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "hidden",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "bindname",
+                               Value = "username"
+                           }
+                        },
+                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
+                        {
+                        }
+                    },
+                    new Portal.Entities.SectionParts.Controls.PageControl
+                    {
+                        Name = "email",
+                        Order = 2,
+                        Type = Portal.Entities.SectionParts.Controls.ControlType.Email,
+                        Options = new List<Portal.Entities.Pages.ShellOption>
+                        {
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "label",
+                               Value = "Email"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "placeholder",
+                               Value = "Enter a email"
                            },
                            new Portal.Entities.Pages.ShellOption
                            {
@@ -88,7 +126,7 @@ namespace LetPortal.Versions.Components
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "bindname",
-                               Value = "name"
+                               Value = "email"
                            }
                         },
                         Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
@@ -97,7 +135,246 @@ namespace LetPortal.Versions.Components
                             {
                                 ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
                                 IsActive = true,
-                                ValidatorMessage = "Please enter a database name"
+                                ValidatorMessage = "Please enter a email"
+                            },
+                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            {
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.MaxLength,
+                                IsActive = true,
+                                ValidatorMessage = "This field requires maximum 250 characters",
+                                ValidatorOption = "250"
+                            },
+                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            {
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Email,
+                                IsActive = true,
+                                ValidatorMessage = "Please enter a correct email format"
+                            }
+                        }
+                    },
+                    new Portal.Entities.SectionParts.Controls.PageControl
+                    {
+                        Name = "isConfirmedEmail",
+                        Order = 3,
+                        Type = Portal.Entities.SectionParts.Controls.ControlType.Checkbox,
+                        Options = new List<Portal.Entities.Pages.ShellOption>
+                        {
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "label",
+                               Value = "Confirmed Email"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "placeholder",
+                               Value = ""
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "disabled",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "hidden",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "bindname",
+                               Value = "isConfirmedEmail"
+                           }
+                        },
+                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
+                        {
+                        }
+                    },
+                    new Portal.Entities.SectionParts.Controls.PageControl
+                    {
+                        Name = "isLockoutEnabled",
+                        Order = 4,
+                        Type = Portal.Entities.SectionParts.Controls.ControlType.Checkbox,
+                        Options = new List<Portal.Entities.Pages.ShellOption>
+                        {
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "label",
+                               Value = "Lockout Enabled"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "placeholder",
+                               Value = ""
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "disabled",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "hidden",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "bindname",
+                               Value = "isLockoutEnabled"
+                           }
+                        },
+                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
+                        {
+                        }
+                    },
+                    new Portal.Entities.SectionParts.Controls.PageControl
+                    {
+                        Name = "lockoutEndDate",
+                        Order = 4,
+                        Type = Portal.Entities.SectionParts.Controls.ControlType.Label,
+                        Options = new List<Portal.Entities.Pages.ShellOption>
+                        {
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "label",
+                               Value = "Lockout End Date"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "placeholder",
+                               Value = ""
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "disabled",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "hidden",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "bindname",
+                               Value = "lockoutEndDate"
+                           }
+                        },
+                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
+                        {
+                        }
+                    },
+                    new Portal.Entities.SectionParts.Controls.PageControl
+                    {
+                        Name = "roles",
+                        Order = 4,
+                        Type = Portal.Entities.SectionParts.Controls.ControlType.AutoComplete,
+                        Options = new List<Portal.Entities.Pages.ShellOption>
+                        {
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "label",
+                               Value = "Roles"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "placeholder",
+                               Value = ""
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "disabled",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "hidden",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "bindname",
+                               Value = "roles"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "multiple",
+                               Value = "true"
+                           }
+                        },
+                        DatasourceOptions = new Portal.Entities.Shared.DatasourceOptions
+                        {
+                            Type = Portal.Entities.Shared.DatasourceControlType.Database,
+                            DatabaseOptions = new Portal.Entities.Shared.DatabaseOptions
+                            {
+                                DatabaseConnectionId = Constants.IdentityDatabaseId,
+                                EntityName = "roles",
+                                Query = versionContext.ConnectionType == Core.Persistences.ConnectionType.MongoDB ?
+                                "{\r\n  \"$query\": {\r\n    \"roles\": [\r\n      {\r\n        \"$project\": {\r\n          \"name\": \"$displayName\",\r\n          \"value\": \"$name\"\r\n        }\r\n      }\r\n    ]\r\n  }\r\n}"
+                                : (versionContext.ConnectionType == Core.Persistences.ConnectionType.MySQL ? "Select `displayName` as name, name as value from `roles`" : "Select \"displayName\" as name, name as value from roles")
+                            }
+                        },
+                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
+                        {
+                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            {
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
+                                IsActive = true,
+                                ValidatorMessage = "Please select at least one role",
+                                ValidatorOption = ""
+                            }
+                        }
+                    }
+                }
+            };
+
+            var registerNewUserFormStandard = new StandardComponent
+            {
+                Id = "5d6222ec3aae6a79ecf035e9",
+                Name = "register-form",
+                DisplayName = "Register Form",
+                LayoutType = PageSectionLayoutType.OneColumn,
+                Controls = new List<Portal.Entities.SectionParts.Controls.PageControl>
+                {
+                    new Portal.Entities.SectionParts.Controls.PageControl
+                    {
+                        Name = "username",
+                        Order = 1,
+                        Type = Portal.Entities.SectionParts.Controls.ControlType.Textbox,
+                        Options = new List<Portal.Entities.Pages.ShellOption>
+                        {
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "label",
+                               Value = "Username"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "placeholder",
+                               Value = "Enter a username"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "disabled",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "hidden",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "bindname",
+                               Value = "username"
+                           }
+                        },
+                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
+                        {
+                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            {
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
+                                IsActive = true,
+                                ValidatorMessage = "Please enter a username"
                             },
                             new Portal.Entities.SectionParts.Controls.PageControlValidator
                             {
@@ -110,8 +387,8 @@ namespace LetPortal.Versions.Components
                             {
                                 ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.MaxLength,
                                 IsActive = true,
-                                ValidatorMessage = "This field requires maximum 250 characters",
-                                ValidatorOption = "250"
+                                ValidatorMessage = "This field requires maximum 50 characters",
+                                ValidatorOption = "50"
                             }
                         },
                         AsyncValidators = new List<Portal.Entities.Components.Controls.PageControlAsyncValidator>
@@ -120,14 +397,14 @@ namespace LetPortal.Versions.Components
                             {
                                 ValidatorName = "uniquename",
                                 IsActive = true,
-                                ValidatorMessage = "This database name is already existed. Please choose another name.",
+                                ValidatorMessage = "This username is already existed. Please choose another name.",
                                 AsyncValidatorOptions = new Portal.Entities.Components.Controls.ControlAsyncValidatorOptions
                                 {
                                     EvaluatedExpression = "response.result == null",
                                     DatabaseOptions = new Portal.Entities.Shared.DatabaseOptions
                                     {
-                                      DatabaseConnectionId = Constants.PortalDatabaseId,
-                                      Query = "{\r\n  \"$query\": {\r\n    \"databases\": [\r\n      {\r\n        \"$match\": {\r\n          \"name\": \"{{data.name}}\"\r\n        }\r\n      }\r\n    ]\r\n  }\r\n}"
+                                      DatabaseConnectionId = Constants.IdentityDatabaseId,
+                                      Query = "{\r\n  \"$query\": {\r\n    \"users\": [\r\n      {\r\n        \"$match\": {\r\n          \"username\": \"{{data.username}}\"\r\n        }\r\n      }\r\n    ]\r\n  }\r\n}"
                                     },
                                     ValidatorType = Portal.Entities.Components.Controls.AsyncValidatorType.DatabaseValidator
                                 }
@@ -136,20 +413,20 @@ namespace LetPortal.Versions.Components
                     },
                     new Portal.Entities.SectionParts.Controls.PageControl
                     {
-                        Name = "dataSource",
+                        Name = "password",
                         Order = 2,
-                        Type = Portal.Entities.SectionParts.Controls.ControlType.Textbox,
+                        Type = Portal.Entities.SectionParts.Controls.ControlType.Password,
                         Options = new List<Portal.Entities.Pages.ShellOption>
                         {
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "label",
-                               Value = "Datasource"
+                               Value = "Password"
                            },
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "placeholder",
-                               Value = "Datasource"
+                               Value = "Enter a password"
                            },
                            new Portal.Entities.Pages.ShellOption
                            {
@@ -164,7 +441,7 @@ namespace LetPortal.Versions.Components
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "bindname",
-                               Value = "dataSource"
+                               Value = "password"
                            }
                         },
                         Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
@@ -173,7 +450,121 @@ namespace LetPortal.Versions.Components
                             {
                                 ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
                                 IsActive = true,
-                                ValidatorMessage = "Please enter a datasource"
+                                ValidatorMessage = "Please enter a password"
+                            },
+                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            {
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.MaxLength,
+                                IsActive = true,
+                                ValidatorMessage = "Password requires maximum 50 characters",
+                                ValidatorOption = "50"
+                            },
+                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            {
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.MinLength,
+                                IsActive = true,
+                                ValidatorMessage = "Password must have at least {{option}} characters",
+                                ValidatorOption = "6"
+                            }
+                        }
+                    },
+                    new Portal.Entities.SectionParts.Controls.PageControl
+                    {
+                        Name = "repassword",
+                        Order = 3,
+                        Type = Portal.Entities.SectionParts.Controls.ControlType.Password,
+                        Options = new List<Portal.Entities.Pages.ShellOption>
+                        {
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "label",
+                               Value = "Confirm password"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "placeholder",
+                               Value = "Enter a confirm password"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "disabled",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "hidden",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "bindname",
+                               Value = "repassword"
+                           }
+                        },
+                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
+                        {
+                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            {
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
+                                IsActive = true,
+                                ValidatorMessage = "Please enter a confirm password"
+                            },
+                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            {
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.MaxLength,
+                                IsActive = true,
+                                ValidatorMessage = "This field requires maximum 50 characters",
+                                ValidatorOption = "50"
+                            },
+                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            {
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.EqualTo,
+                                IsActive = true,
+                                ValidatorMessage = "This field must be equal to {{option}}",
+                                ValidatorOption = "password"
+                            }
+                        }
+                    },
+                    new Portal.Entities.SectionParts.Controls.PageControl
+                    {
+                        Name = "email",
+                        Order = 4,
+                        Type = Portal.Entities.SectionParts.Controls.ControlType.Email,
+                        Options = new List<Portal.Entities.Pages.ShellOption>
+                        {
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "label",
+                               Value = "Email"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "placeholder",
+                               Value = "Enter a email"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "disabled",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "hidden",
+                               Value = "false"
+                           },
+                           new Portal.Entities.Pages.ShellOption
+                           {
+                               Key = "bindname",
+                               Value = "email"
+                           }
+                        },
+                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
+                        {
+                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            {
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
+                                IsActive = true,
+                                ValidatorMessage = "Please enter a email"
                             },
                             new Portal.Entities.SectionParts.Controls.PageControlValidator
                             {
@@ -181,117 +572,42 @@ namespace LetPortal.Versions.Components
                                 IsActive = true,
                                 ValidatorMessage = "This field requires maximum 250 characters",
                                 ValidatorOption = "250"
-                            }
-                        }
-                    },
-                    new Portal.Entities.SectionParts.Controls.PageControl
-                    {
-                        Name = "connectionString",
-                        Order = 3,
-                        Type = Portal.Entities.SectionParts.Controls.ControlType.Textbox,
-                        Options = new List<Portal.Entities.Pages.ShellOption>
-                        {
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "label",
-                               Value = "Connection String"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "placeholder",
-                               Value = "Connection String"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "disabled",
-                               Value = "false"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "hidden",
-                               Value = "false"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "bindname",
-                               Value = "connectionString"
-                           }
-                        },
-                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
-                        {
-                            new Portal.Entities.SectionParts.Controls.PageControlValidator
-                            {
-                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
-                                IsActive = true,
-                                ValidatorMessage = "Please enter a connection string"
                             },
                             new Portal.Entities.SectionParts.Controls.PageControlValidator
                             {
-                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.MaxLength,
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Email,
                                 IsActive = true,
-                                ValidatorMessage = "This field requires maximum 500 characters",
-                                ValidatorOption = "500"
-                            }
-                        }
-                    },
-                    new Portal.Entities.SectionParts.Controls.PageControl
-                    {
-                        Name = "databaseConnectionType",
-                        Order = 4,
-                        Type = Portal.Entities.SectionParts.Controls.ControlType.Select,
-                        Options = new List<Portal.Entities.Pages.ShellOption>
-                        {
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "label",
-                               Value = "Connection Type"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "placeholder",
-                               Value = ""
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "disabled",
-                               Value = "false"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "hidden",
-                               Value = "false"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "bindname",
-                               Value = "databaseConnectionType"
-                           }
-                        },
-                        DatasourceOptions = new Portal.Entities.Shared.DatasourceOptions
-                        {
-                            DatasourceStaticOptions = new Portal.Entities.Shared.DatasourceStaticOptions
-                            {
-                                JsonResource = "[{\"name\":\"MongoDB\",\"value\":\"mongodb\"},{\"name\":\"SQL Server\",\"value\":\"sqlserver\"}, {\"name\":\"PostgreSQL\",\"value\":\"postgresql\"}, {\"name\":\"MySQL\",\"value\":\"mysql\"}]"
+                                ValidatorMessage = "This field must be email format"
                             }
                         },
-                         Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
+                         AsyncValidators = new List<Portal.Entities.Components.Controls.PageControlAsyncValidator>
                         {
-                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            new Portal.Entities.Components.Controls.PageControlAsyncValidator
                             {
-                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
+                                ValidatorName = "uniqueemail",
                                 IsActive = true,
-                                ValidatorMessage = "Please choose one connection type"
+                                ValidatorMessage = "This email is already existed. Please choose another email.",
+                                AsyncValidatorOptions = new Portal.Entities.Components.Controls.ControlAsyncValidatorOptions
+                                {
+                                    EvaluatedExpression = "response.result == null",
+                                    DatabaseOptions = new Portal.Entities.Shared.DatabaseOptions
+                                    {
+                                      DatabaseConnectionId = Constants.IdentityDatabaseId,
+                                      Query = "{\r\n  \"$query\": {\r\n    \"users\": [\r\n      {\r\n        \"$match\": {\r\n          \"email\": \"{{data.emaill}}\"\r\n        }\r\n      }\r\n    ]\r\n  }\r\n}"
+                                    },
+                                    ValidatorType = Portal.Entities.Components.Controls.AsyncValidatorType.DatabaseValidator
+                                }
                             }
                         }
                     }
                 }
             };
 
-            var appStandard = new StandardComponent
+            var roleStandard = new StandardComponent
             {
-                Id = "5d3836194d8fa90874135d68",
-                Name = "app-form",
-                DisplayName = "App Form",
+                Id = "5d63423dbaac7d4790b7d300",
+                Name = "role-form",
+                DisplayName = "Role Form",
                 LayoutType = PageSectionLayoutType.OneColumn,
                 Controls = new List<Portal.Entities.SectionParts.Controls.PageControl>
                 {
@@ -305,12 +621,12 @@ namespace LetPortal.Versions.Components
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "label",
-                               Value = "Id"
+                               Value = "id"
                            },
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "placeholder",
-                               Value = ""
+                               Value = "id"
                            },
                            new Portal.Entities.Pages.ShellOption
                            {
@@ -339,12 +655,12 @@ namespace LetPortal.Versions.Components
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "label",
-                               Value = "Name"
+                               Value = "Role Name"
                            },
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "placeholder",
-                               Value = "Name"
+                               Value = "Enter a role name"
                            },
                            new Portal.Entities.Pages.ShellOption
                            {
@@ -368,21 +684,21 @@ namespace LetPortal.Versions.Components
                             {
                                 ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
                                 IsActive = true,
-                                ValidatorMessage = "Please enter a app name"
-                            },
-                            new Portal.Entities.SectionParts.Controls.PageControlValidator
-                            {
-                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Regex,
-                                IsActive = true,
-                                ValidatorMessage = "Please enter only characters and numberic",
-                                ValidatorOption = "[A-Za-z0-9]*"
+                                ValidatorMessage = "Please enter a name"
                             },
                             new Portal.Entities.SectionParts.Controls.PageControlValidator
                             {
                                 ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.MaxLength,
                                 IsActive = true,
-                                ValidatorMessage = "This field requires maximum 250 characters",
-                                ValidatorOption = "250"
+                                ValidatorMessage = "This field requires maximum 50 characters",
+                                ValidatorOption = "50"
+                            },
+                            new Portal.Entities.SectionParts.Controls.PageControlValidator
+                            {
+                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Regex,
+                                IsActive = true,
+                                ValidatorMessage = "Please enter only characters",
+                                ValidatorOption = "[A-Za-z]*"
                             }
                         },
                         AsyncValidators = new List<Portal.Entities.Components.Controls.PageControlAsyncValidator>
@@ -391,21 +707,21 @@ namespace LetPortal.Versions.Components
                             {
                                 ValidatorName = "uniquename",
                                 IsActive = true,
-                                ValidatorMessage = "This app name is already existed. Please choose another name.",
+                                ValidatorMessage = "This role is already existed. Please choose another name.",
                                 AsyncValidatorOptions = new Portal.Entities.Components.Controls.ControlAsyncValidatorOptions
                                 {
                                     EvaluatedExpression = "response.result == null",
                                     DatabaseOptions = new Portal.Entities.Shared.DatabaseOptions
                                     {
-                                      DatabaseConnectionId = Constants.PortalDatabaseId,
-                                      Query = "{\r\n  \"$query\": {\r\n    \"apps\": [\r\n      {\r\n        \"$match\": {\r\n          \"name\": \"{{data.name}}\"\r\n        }\r\n      }\r\n    ]\r\n  }\r\n}"
+                                      DatabaseConnectionId = Constants.IdentityDatabaseId,
+                                      Query = "{\r\n  \"$query\": {\r\n    \"roles\": [\r\n      {\r\n        \"$match\": {\r\n          \"name\": \"{{data.name}}\"\r\n        }\r\n      }\r\n    ]\r\n  }\r\n}"
                                     },
                                     ValidatorType = Portal.Entities.Components.Controls.AsyncValidatorType.DatabaseValidator
                                 }
                             }
                         }
                     },
-                     new Portal.Entities.SectionParts.Controls.PageControl
+                    new Portal.Entities.SectionParts.Controls.PageControl
                     {
                         Name = "displayName",
                         Order = 2,
@@ -415,12 +731,12 @@ namespace LetPortal.Versions.Components
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "label",
-                               Value = "Display Name"
+                               Value = "Role Display Name"
                            },
                            new Portal.Entities.Pages.ShellOption
                            {
                                Key = "placeholder",
-                               Value = "Display Name"
+                               Value = "Enter a display name"
                            },
                            new Portal.Entities.Pages.ShellOption
                            {
@@ -450,161 +766,21 @@ namespace LetPortal.Versions.Components
                             {
                                 ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.MaxLength,
                                 IsActive = true,
-                                ValidatorMessage = "This field requires maximum 250 characters",
-                                ValidatorOption = "250"
-                            }
-                        }
-                    },
-                     new Portal.Entities.SectionParts.Controls.PageControl
-                    {
-                        Name = "defaultUrl",
-                        Order = 3,
-                        Type = Portal.Entities.SectionParts.Controls.ControlType.Textbox,
-                        Options = new List<Portal.Entities.Pages.ShellOption>
-                        {
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "label",
-                               Value = "Default Url"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "placeholder",
-                               Value = "Default Url"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "disabled",
-                               Value = "false"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "hidden",
-                               Value = "false"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "bindname",
-                               Value = "defaultUrl"
-                           }
-                        },
-                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
-                        {
-                            new Portal.Entities.SectionParts.Controls.PageControlValidator
-                            {
-                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
-                                IsActive = true,
-                                ValidatorMessage = "Please enter a default url"
-                            },
-                            new Portal.Entities.SectionParts.Controls.PageControlValidator
-                            {
-                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.MaxLength,
-                                IsActive = true,
-                                ValidatorMessage = "This field requires maximum 250 characters",
-                                ValidatorOption = "250"
-                            }
-                        }
-                    },
-                     new Portal.Entities.SectionParts.Controls.PageControl
-                    {
-                        Name = "currentVersionNumber",
-                        Order = 4,
-                        Type = Portal.Entities.SectionParts.Controls.ControlType.Textbox,
-                        Options = new List<Portal.Entities.Pages.ShellOption>
-                        {
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "label",
-                               Value = "Version"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "placeholder",
-                               Value = "App Version"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "disabled",
-                               Value = "false"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "hidden",
-                               Value = "false"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "bindname",
-                               Value = "currentVersionNumber"
-                           }
-                        },
-                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
-                        {
-                            new Portal.Entities.SectionParts.Controls.PageControlValidator
-                            {
-                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
-                                IsActive = true,
-                                ValidatorMessage = "Please enter a version number"
-                            },
-                            new Portal.Entities.SectionParts.Controls.PageControlValidator
-                            {
-                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.MaxLength,
-                                IsActive = true,
                                 ValidatorMessage = "This field requires maximum 50 characters",
                                 ValidatorOption = "50"
-                            }
-                        }
-                    },
-                     new Portal.Entities.SectionParts.Controls.PageControl
-                    {
-                        Name = "logo",
-                        Order = 4,
-                        Type = Portal.Entities.SectionParts.Controls.ControlType.IconPicker,
-                        Options = new List<Portal.Entities.Pages.ShellOption>
-                        {
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "label",
-                               Value = "Icon"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "placeholder",
-                               Value = "Icon"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "disabled",
-                               Value = "false"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "hidden",
-                               Value = "false"
-                           },
-                           new Portal.Entities.Pages.ShellOption
-                           {
-                               Key = "bindname",
-                               Value = "logo"
-                           }
-                        },
-                        Validators = new List<Portal.Entities.SectionParts.Controls.PageControlValidator>
-                        {
-                            new Portal.Entities.SectionParts.Controls.PageControlValidator
-                            {
-                                ValidatorType = Portal.Entities.SectionParts.Controls.ValidatorType.Required,
-                                IsActive = true,
-                                ValidatorMessage = "Please choose one icon"
                             }
                         }
                     }
                 }
             };
 
-            VersionUtils.GenerateControlEvents(databaseStandard);
-            VersionUtils.GenerateControlEvents(appStandard);
-            versionContext.InsertData(databaseStandard);
-            versionContext.InsertData(appStandard);
+            VersionUtils.GenerateControlEvents(userFormStandard);
+            VersionUtils.GenerateControlEvents(registerNewUserFormStandard);
+            VersionUtils.GenerateControlEvents(roleStandard);
+
+            versionContext.InsertData(userFormStandard);
+            versionContext.InsertData(registerNewUserFormStandard);
+            versionContext.InsertData(roleStandard);
         }
     }
 }
