+++ src/web-apis/LetPortal.Versions/Databases/Database_0_1_0.cs
@@ -5,9 +5,9 @@ using LetPortal.Portal.Entities.Databases;
 
 namespace LetPortal.Versions.Databases
 {
-    public class Database_0_0_1 : IPortalVersion
+    public class Database_0_1_0 : IPortalVersion
     {
-        public string VersionNumber => "0.0.1";
+        public string VersionNumber => "0.1.0";
 
         public void Downgrade(IVersionContext versionContext)
         {
