+++ src/web-apis/LetPortal.Versions/Datasources/Datasource_0_1_0.cs
@@ -3,9 +3,9 @@ using LetPortal.Portal.Entities.Datasources;
 
 namespace LetPortal.Versions.Datasources
 {
-    public class Datasource_0_0_1 : IPortalVersion
+    public class Datasource_0_1_0 : IPortalVersion
     {
-        public string VersionNumber => "0.0.1";
+        public string VersionNumber => "0.1.0";
 
         public void Downgrade(IVersionContext versionContext)
         {
