+++ src/web-apis/LetPortal.WebApis/Controllers/ChartsController.cs
@@ -142,5 +142,11 @@ namespace LetPortal.PortalApis.Controllers
             return BadRequest();
         }
 
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> Delete(string id)
+        {
+            await _chartRepository.DeleteAsync(id);
+            return Ok();
+        }
     }
 }
