+++ test/IntegrationTests/LetPortal.Tests.ITs.Portal/Repositories/PageRepositoryTests.cs
@@ -175,11 +175,6 @@ namespace LetPortal.Tests.ITs.Portal.Repositories
                         IsRequiredValidation = true,
                         ButtonOptions = new ButtonOptions
                         {
-                            ConfirmationOptions = new ConfirmationOptions
-                            {
-                                IsEnable = true,
-                                ConfirmationText = "Are you sure to create a database?"
-                            },
                             ActionCommandOptions = new ActionCommandOptions
                             {
                                 ActionType = ActionType.ExecuteDatabase,
@@ -198,6 +193,11 @@ namespace LetPortal.Tests.ITs.Portal.Repositories
                                 {
                                     CompleteMessage = "Insert new database successfully!",
                                     FailedMessage = "Oops! Something went wrong, please try again!"
+                                },
+                                ConfirmationOptions = new ConfirmationOptions
+                                {
+                                    IsEnable = true,
+                                    ConfirmationText = "Are you sure to create a database?"
                                 }
                             },
                             RouteOptions = new RouteOptions
